requirements.txt
django>=4.2,<5.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
psycopg2-binary>=2.9
django-filter>=23.0
gunicorn>=21.0
redis>=5.0
celery>=5.3

# Desarrollo y tests
pytest-django>=4.7
pytest>=7.4
ruff>=0.1
